rustorch::training::trainer

Trait TrainingDataLoader

pub trait TrainingDataLoader<T: Float> {
    // Required methods
    fn reset(&mut self);
    fn next_batch(&mut self) -> Option<(Tensor<T>, Tensor<T>)>;
    fn is_empty(&self) -> bool;
}

Training data loader trait for providing batches of (input, target) pairs 訓練データローダートレイト

Required Methods

fn reset(&mut self)

Reset the data loader for a new epoch 新しいエポックのためにデータローダーをリセット

fn next_batch(&mut self) -> Option<(Tensor<T>, Tensor<T>)>

Get next batch of (input, target) tensor pairs 次の(入力, ターゲット)テンソルペアのバッチを取得

fn is_empty(&self) -> bool

Check if the data loader is exhausted データローダーが枯渇したかチェック