rustorch::vision::pipeline

Module predicates

Predefined common predicates for conditional transforms 条件付き変換用の事前定義された一般的述語

Functions

channels_eq

Only apply transform to images with specific number of channels 特定のチャンネル数の画像にのみ変換を適用

format_is

Only apply transform to images with specific format 特定の形式の画像にのみ変換を適用

max_size

Only apply transform to images smaller than specified dimensions 指定された次元より小さい画像にのみ変換を適用

min_size

Only apply transform to images larger than specified dimensions 指定された次元より大きい画像にのみ変換を適用

probability

Apply transform with specified probability 指定された確率で変換を適用